Update fibhash src files

+++ template/src/S01E02-FibHash.sol
@@ -7,7 +7,7 @@ contract FibHash is IFibHash {
     /// @dev Implementation of the fibonacci hash algorithm
     /// @param x The number to hash
     /// @param k The size of the resulting hash table: 2^k
-    function fibhash(uint64 x, uint64 k) public pure returns (uint64) {
+    function fibhash(uint256 x, uint8 k) public pure returns (uint256) {
         return 0;
     }
 }
